Config bringup files
